blog/scripts/requirements.txt
markdown>=3.4.1
PyYAML>=6.0
pathlib>=1.0.1 